removed identity changes into the identity semantics section that was there
In the yyc version there are two places that identity semantics is discussed: it said that = is special 2-place predicate with a fixed interpretation in the formalizations part and had a separate semantics section. The result of my changes yesterday and today is that the early talk about special 2-place predicate is removed and the semantics for FOL with identity is updated:
Expanded examples in sentences and worked through example of semantics with quantifiers and identity.